Simplify model-extension.
