+++ test/provider/create-model.spec.js
@@ -4,10 +4,10 @@ const sinon = require('sinon')
 require('should-sinon')
 const _ = require('lodash')
 const providerMock = require('../fixtures/fake-provider')
-const ExtendedModel = require('../../src/provider/extended-model')
+const createModel = require('../../src/provider/create-model')
 const koopMock = { test: 'value' }
 
-describe('Tests for extended-model', function () {
+describe('Tests for create-model', function () {
   describe('createKey', function () {
     it('should create cache-key as provider name', () => {
       const retrieveSpy = sinon.spy(function (key, queryParams, callback) {
@@ -16,7 +16,7 @@ describe('Tests for extended-model', function () {
       const pullSpy = sinon.spy()
 
       // create a model with mocked cache "retrieve" function
-      const model = new ExtendedModel({ ProviderModel: providerMock.Model, koop: koopMock }, {
+      const model = createModel({ ProviderModel: providerMock.Model, koop: koopMock }, {
         cache: {
           retrieve: retrieveSpy
         }
@@ -42,7 +42,7 @@ describe('Tests for extended-model', function () {
       const pullSpy = sinon.spy()
 
       // create a model with mocked cache "retrieve" function
-      const model = new ExtendedModel({ ProviderModel: providerMock.Model, koop: koopMock }, {
+      const model = createModel({ ProviderModel: providerMock.Model, koop: koopMock }, {
         cache: {
           retrieve: retrieveSpy
         }
@@ -67,7 +67,7 @@ describe('Tests for extended-model', function () {
       const pullSpy = sinon.spy()
 
       // create a model with mocked cache "retrieve" function
-      const model = new ExtendedModel({ ProviderModel: providerMock.Model, koop: koopMock }, {
+      const model = createModel({ ProviderModel: providerMock.Model, koop: koopMock }, {
         cache: {
           retrieve: retrieveSpy
         }
@@ -99,7 +99,7 @@ describe('Tests for extended-model', function () {
         callback(null, { metadata: {} })
       })
 
-      const model = new ExtendedModel({ ProviderModel: providerMock.Model, koop: koopMock }, {
+      const model = createModel({ ProviderModel: providerMock.Model, koop: koopMock }, {
         cache: { retrieve: (key, query, callback) => { callback(new Error('no cache')) } }
       })
       model.before = beforeSpy
@@ -133,7 +133,7 @@ describe('Tests for extended-model', function () {
         callback(null, data)
       })
       const pullCallbackSpy = sinon.spy(function (err, data) {})
-      const model = new ExtendedModel({ ProviderModel: providerMock.Model, koop: koopMock }, {
+      const model = createModel({ ProviderModel: providerMock.Model, koop: koopMock }, {
         cache: { retrieve: (key, query, callback) => { callback(new Error('no cache')) } }
       })
       model.getData = getDataSpy
@@ -163,7 +163,7 @@ describe('Tests for extended-model', function () {
   describe('fetch layer metadata', function () {
     afterEach(function () {
       // reset the getLayer() function to default
-      ExtendedModel.prototype.getLayer = undefined
+      createModel.prototype.getLayer = undefined
     })
 
     it('should throw an error if the getLayer() function is not implemented', function () {
@@ -173,7 +173,7 @@ describe('Tests for extended-model', function () {
       const callbackSpy = sinon.spy()
 
       // create a model with mocked cache "retrieve" function
-      const model = new ExtendedModel({ ProviderModel: providerMock.Model, koop: koopMock }, {
+      const model = createModel({ ProviderModel: providerMock.Model, koop: koopMock }, {
         cache: {
           retrieve: retrieveSpy
         }
@@ -196,10 +196,10 @@ describe('Tests for extended-model', function () {
       })
       const callbackSpy = sinon.spy()
 
-      ExtendedModel.prototype.getLayer = getLayerSpy
+      createModel.prototype.getLayer = getLayerSpy
 
       // create a model with mocked cache "retrieve" function
-      const model = new ExtendedModel({ ProviderModel: providerMock.Model, koop: koopMock }, {
+      const model = createModel({ ProviderModel: providerMock.Model, koop: koopMock }, {
         cache: {
           retrieve: retrieveSpy
         }
@@ -229,7 +229,7 @@ describe('Tests for extended-model', function () {
       const callbackSpy = sinon.spy()
 
       // create a model with mocked cache "retrieve" function
-      const model = new ExtendedModel({ ProviderModel: providerMock.Model, koop: koopMock }, {
+      const model = createModel({ ProviderModel: providerMock.Model, koop: koopMock }, {
         cache: {
           retrieve: retrieveSpy
         }
@@ -251,7 +251,7 @@ describe('Tests for extended-model', function () {
   describe('fetch catalog metadata', function () {
     afterEach(function () {
       // reset the getCatalog() function to default
-      ExtendedModel.prototype.getCatalog = undefined
+      createModel.prototype.getCatalog = undefined
     })
 
     it('should throw an error if the getCatalog() function is not implemented', function () {
@@ -261,7 +261,7 @@ describe('Tests for extended-model', function () {
       const callbackSpy = sinon.spy()
 
       // create a model with mocked cache "retrieve" function
-      const model = new ExtendedModel({ ProviderModel: providerMock.Model, koop: koopMock }, {
+      const model = createModel({ ProviderModel: providerMock.Model, koop: koopMock }, {
         cache: {
           retrieve: retrieveSpy
         }
@@ -285,7 +285,7 @@ describe('Tests for extended-model', function () {
       const callbackSpy = sinon.spy()
 
       // create a model with mocked cache "retrieve" function
-      const model = new ExtendedModel({ ProviderModel: providerMock.Model, koop: koopMock }, {
+      const model = createModel({ ProviderModel: providerMock.Model, koop: koopMock }, {
         cache: {
           retrieve: retrieveSpy
         }
@@ -317,7 +317,7 @@ describe('Tests for extended-model', function () {
       const callbackSpy = sinon.spy()
 
       // create a model with mocked cache "retrieve" function
-      const model = new ExtendedModel({ ProviderModel: providerMock.Model, koop: koopMock }, {
+      const model = createModel({ ProviderModel: providerMock.Model, koop: koopMock }, {
         cache: {
           retrieve: retrieveSpy
         }
